+++ routes/api.php
@@ -84,24 +84,24 @@ Route::prefix('v1')->name('api.v1.')->group(function () {
         Route::get('rooms/{room}/member', [RoomMemberController::class,'index'])->name('rooms.member.get')->middleware('can:viewMembers,room');
         Route::post('rooms/{room}/member', [RoomMemberController::class,'store'])->name('rooms.member.add')->middleware('can:manageMembers,room');
         Route::put('rooms/{room}/member/{user}', [RoomMemberController::class,'update'])->name('rooms.member.update')->middleware('can:manageMembers,room');
-        Route::delete('rooms/{room}/member/{user}', [RoomMemberController::class,'destroy'])->name('rooms.member.remove')->middleware('can:manageMembers,room');
+        Route::delete('rooms/{room}/member/{user}', [RoomMemberController::class,'destroy'])->name('rooms.member.destroy')->middleware('can:manageMembers,room');
 
         // Membership users for mass update & delete
         Route::put('rooms/{room}/member', [RoomMemberController::class, 'bulkUpdate'])->name('rooms.member.bulkUpdate')->middleware('can:manageMembers,room');
-        Route::delete('rooms/{room}/member', [RoomMemberController::class, 'bulkDestroy'])->name('rooms.member.bulkRemove')->middleware('can:manageMembers,room');
+        Route::delete('rooms/{room}/member', [RoomMemberController::class, 'bulkDestroy'])->name('rooms.member.bulkDestroy')->middleware('can:manageMembers,room');
 
         // Personalized room tokens
         Route::get('rooms/{room}/tokens', [RoomTokenController::class, 'index'])->name('rooms.tokens.get')->middleware('can:viewTokens,room');
         Route::post('rooms/{room}/tokens', [RoomTokenController::class, 'store'])->name('rooms.tokens.add')->middleware('can:manageTokens,room');
         Route::put('rooms/{room}/tokens/{token}', [RoomTokenController::class, 'update'])->name('rooms.tokens.update')->middleware('can:manageTokens,room');
-        Route::delete('rooms/{room}/tokens/{token}', [RoomTokenController::class, 'destroy'])->name('rooms.tokens.remove')->middleware('can:manageTokens,room');
+        Route::delete('rooms/{room}/tokens/{token}', [RoomTokenController::class, 'destroy'])->name('rooms.tokens.destroy')->middleware('can:manageTokens,room');
 
         // File operations
         Route::middleware('can:manageFiles,room')->group(function () {
             Route::post('rooms/{room}/files', [RoomFileController::class,'store'])->name('rooms.files.add');
 
             Route::put('rooms/{room}/files/{file}', [RoomFileController::class,'update'])->name('rooms.files.update');
-            Route::delete('rooms/{room}/files/{file}', [RoomFileController::class,'destroy'])->name('rooms.files.remove');
+            Route::delete('rooms/{room}/files/{file}', [RoomFileController::class,'destroy'])->name('rooms.files.destroy');
         });
 
         Route::get('users/search', [UserController::class,'search'])->name('users.search');
@@ -119,7 +119,7 @@ Route::prefix('v1')->name('api.v1.')->group(function () {
 
 
         Route::get('sessions', [SessionController::class,'index'])->name('sessions.index');
-        Route::delete('sessions', [SessionController::class,'destroy'])->name('sessions.delete');
+        Route::delete('sessions', [SessionController::class,'destroy'])->name('sessions.destroy');
 
         Route::post('servers/check', [ServerController::class,'check'])->name('servers.check')->middleware('can:viewAny,App\Models\Server');
         Route::get('servers/{server}/panic', [ServerController::class,'panic'])->name('servers.panic')->middleware('can:update,server');