+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Support to customize trusted proxies ([#305],[#306])
+
 ## [v2.0.0-alpha.5] - 2023-03-09
 ### Added
 - Confirmation for password and email change ([#277],[#304])
@@ -429,6 +432,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#297]: https://github.com/THM-Health/PILOS/issues/297
 [#298]: https://github.com/THM-Health/PILOS/issues/298
 [#304]: https://github.com/THM-Health/PILOS/issues/304
+[#305]: https://github.com/THM-Health/PILOS/issues/305
+[#306]: https://github.com/THM-Health/PILOS/issues/306
 
 [unreleased]: https://github.com/THM-Health/PILOS/compare/v2.0.0-alpha.5...HEAD
 [1.0.0]: https://github.com/THM-Health/PILOS/releases/tag/v1.0.0